+++ R/save.R
@@ -1,9 +1,14 @@
-#' S3 generics to save returned values from functions in QntMap package.
+#' Save objects created by qntmap package.
 #' @return `invisible(x)`
-#' @param x object to save
-#' @param nm name of object to save
-#' @param saving `TRUE` or `FALSE` to save result file(s).
-#' @param ... other arguments passed to class methods
+#' @param x      An object to save.
+#' @param nm     A name of object to be saved as.
+#' @param saving `TRUE` (default) or `FALSE` to save.
+#' @param ...    Other arguments passed to class methods
+#' @importFrom data.table fwrite
+#' @importFrom ggplot2    ggsave labs
+#' @importFrom pipeR      %>>%
+#' @importFrom png        writePNG
+#' @importFrom purrr      walk2
 #' @export
 save4qm <- function(x, nm = '', saving = TRUE, ...) {
   if(!saving) return(invisible(x))
@@ -13,8 +18,7 @@ save4qm <- function(x, nm = '', saving = TRUE, ...) {
 
 #' @rdname save4qm
 #' @section `save4qm.default`: 
-#'   A default method.
-#'   This is a wrapper of `saveRDS` together with invisible return of `x`.
+#'   A default method saves an object as a binary RDS file.
 save4qm.default <- function(x, nm, saving, ...) {
   saveRDS(x, nm)
   invisible(x)
@@ -22,11 +26,7 @@ save4qm.default <- function(x, nm, saving, ...) {
 
 #' @rdname save4qm
 #' @section `save4qm.data.frame`:
-#'   A method for `data.frame` class. 
-#'   This is a wrapper of `data.table::fwrite` 
-#'   together with invisible return of `x`.
-#' 
-#' @importFrom data.table fwrite
+#'   A data frame object is saved as a csv. 
 save4qm.data.frame <- function(x, nm, saving, ...) {
   fwrite(x, nm)
   invisible(x)
@@ -34,13 +34,9 @@ save4qm.data.frame <- function(x, nm, saving, ...) {
 
 #' @rdname save4qm
 #' @section `save4qm.qm_cluster`: 
-#' A method for `qm_cluster` class returned by 
-#' [`cluster_xmap()`] or [`group_cluster()`].
-#' Object is saved as RDS file and png file.
-#' The latter shows distribution of phases among a mapped area.
-#' @importFrom png writePNG
-#' @importFrom grDevices dev.copy dev.off png
-#' @importFrom graphics pie
+#'  A value returned by [`cluster_xmap()`] or [`group_cluster()`] is
+#'  saved as binary RDS file, dot-by-dot png file, and as svg file with a legend.
+#'  The png and svg files show distribution of phases among a mapped area.
 save4qm.qm_cluster <- function(x, nm, saving, ...) {
   #setting for output
   dir_out <- paste0(x$dir_map, '/clustering')
@@ -51,36 +47,30 @@ save4qm.qm_cluster <- function(x, nm, saving, ...) {
       paste(x$elements, collapse ='')
     )
 
-  #save modal map
-  writePNG(
-    image = array(
-      mycolors(n = k, dec = TRUE)[x$ytehat, ] / 255,
-      dim = c(x$dims, 3)
-    ),
-    target = paste0(nm, "_map.png")
-  )
-
-  #save result of classification
-  saveRDS(x, paste0(nm, "_result.RDS"))
-
-  #save legend
-  pie(
-    rep(1, k), 
-    labels = paste(1:k, colnames(x$membership), sep='.'), 
-    col = mycolors(n = k)
-  )
-  dev.copy(png, paste0(nm, "_legend.png"))
-  dev.off()
+  .cls <- as.factor(x$cluster)
+  .img <- as_img(lookup$discrete(.cls), x$dims[1], x$dims[2])
   
+  # Save as binary
+  saveRDS(x, paste0(nm, "_result.RDS"))
+  
+  # Save as dot-by-dot png
+  writePNG(image = .img, target = paste0(nm, "_map.png"))
+
+  # Save plot with legend as svg
+  A4 <- c(210, 297)[order(x$dims)] # 1st = H and 2nd = W in millimeters
+  ggsave(
+    paste0(nm, "_legend.svg"), 
+    gg_img(.img, zlim = levels(.cls)) + labs(y = "Rows", x = "Columns"),
+    height = A4[[1]], width = A4[[2]], units = "mm"
+  )
+
   invisible(x)
 }
 
 #' @rdname save4qm
 #' @section `save4qm.qntmap`: 
-#' A method for `qntmap` class returned by `qntmap()` or `quantify()`
-#' 
-#' @importFrom data.table fwrite
-#' @importFrom purrr walk2
+#'  A value returned by `qntmap()` or `quantify()` is saved as
+#'  a binary RDS file, and csv files.
 save4qm.qntmap <- function(x, nm, saving, ...) {
   saveRDS(x, file.path(nm, 'qntmap.RDS'))
   unlist(x, recursive = FALSE) %>>%